+++ packages/backend/src/types.ts
@@ -1,6 +1,6 @@
-import { Explore } from 'common';
+import { Explore, ExploreError } from 'common';
 
 export interface ProjectAdapter {
-    compileAllExplores(): Promise<Explore[]>;
+    compileAllExplores(): Promise<(Explore | ExploreError)[]>;
     runQuery(sql: string): Promise<Record<string, any>>;
 }